requirements.txt
psutil
pyserial
pystray
Pillow
pygetwindow
winshell
pywin32
